remove harvest_plugin_extended
